Move to domain layer
