+++ source/Messaging.Application/OutgoingMessages/DocumentType.cs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Messaging.Domain.SeedWork;
+
 namespace Messaging.Application.OutgoingMessages;
 
 public class DocumentType : EnumerationType
